+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-08-11 22:44:23
-// Commit: perf(test): optimize memory consumption in test list viewport
+// Updated at: 2026-08-13 09:01:40
+// Commit: test(test): implement widget integration tests for test
 
 class testComponentState {
-  final String commitRef = '3554';
-  final String lastMessage = 'perf(test): optimize memory consumption in test list viewport';
-  final DateTime timestamp = DateTime.parse('2026-08-11T22:44:23+05:00');
+  final String commitRef = '3558';
+  final String lastMessage = 'test(test): implement widget integration tests for test';
+  final DateTime timestamp = DateTime.parse('2026-08-13T09:01:40+05:00');
 }